- New required Python version is 3.8. (Due to the new walrus operator 🥰)
- Removed the ugly _translate_date() method. Replaced by the
dateparser library.

- Fixed the bug in _parseFileProperties() where the parsing failed for an item
with only one 'il_ItemProperty'

- Added a few tests for _parseFileProperties().

- Refactoring of scanFolder() into scanMediaContainer() and scanContainerList().

- Better scanning for videos via the 'ilPageVideo' css-class instead of
'ilc_media_caption_MediaCaption'. Fixes the bug for videos without a
caption.
